requirements.txt
av==11.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
filelock==3.9.0
fonttools==4.45.1
fsspec==2023.4.0
fvcore==0.1.5.post20221221
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parameterized==0.9.0
Pillow==10.1.0
portalocker==2.8.2
proglog==0.1.10
pyparsing==3.1.1
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==2.1.0
yacs==0.1.8
zipp==3.17.0
